tests: move HugePagesConfig to utils_hugepages
HugePagesConfig describes the supported huge-page modes and is shared
by the test framework and its tests. Move it from framework.microvm to
a dedicated utils_hugepages module, following the utils_<topic>
convention, and update its import sites.

Pure move: no functional change.
