+++ tests/integration_tests/performance/test_balloon.py
@@ -9,12 +9,8 @@ import time
 import pytest
 
 from framework.artifacts import GUEST_KERNEL_DEFAULT, pin_guest_kernel
-from framework.microvm import HugePagesConfig
-from framework.utils import (
-    get_stable_rss_mem,
-    supports_hugetlbfs_discard,
-    track_cpu_utilization,
-)
+from framework.utils import get_stable_rss_mem, track_cpu_utilization
+from framework.utils_hugepages import HugePagesConfig, supports_hugetlbfs_discard
 
 # Every test in this module exercises all huge_pages variants.
 pytestmark = pytest.mark.parametrize(
