+++ tests/integration_tests/performance/test_hotplug_memory.py
@@ -13,13 +13,10 @@ from tenacity import Retrying, retry_if_exception_type, stop_after_delay, wait_f
 
 from framework.artifacts import GUEST_KERNEL_DEFAULT, pin_guest_kernel
 from framework.guest_stats import MeminfoGuest
-from framework.microvm import HugePagesConfig, SnapshotType
+from framework.microvm import SnapshotType
 from framework.properties import global_props
-from framework.utils import (
-    get_resident_memory,
-    make_guest_dirty_memory,
-    supports_hugetlbfs_discard,
-)
+from framework.utils import get_resident_memory, make_guest_dirty_memory
+from framework.utils_hugepages import HugePagesConfig, supports_hugetlbfs_discard
 
 MEMHP_BOOTARGS = "console=ttyS0 reboot=k panic=1 memhp_default_state=online_movable"
 DEFAULT_CONFIG = {"total_size_mib": 1024, "slot_size_mib": 128, "block_size_mib": 2}
